Rename ObjectBinderHelpers to PocoTableEntityConverter (fixes #177).
