+++ test/Microsoft.Azure.WebJobs.Host.UnitTests/WebJobs.Host.UnitTests.csproj
@@ -145,7 +145,7 @@
     <Compile Include="Protocols\JsonTypeNameAttributeTests.cs" />
     <Compile Include="Timers\LinearSpeedupStrategyTests.cs" />
     <Compile Include="Indexers\NameResolverTests.cs" />
-    <Compile Include="ObjectBinderHelpersTests.cs" />
+    <Compile Include="Tables\PocoTableEntityConverterTests.cs" />
     <Compile Include="Protocols\PolymorphicJsonConverterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PublicSurfaceTests.cs" />
